helm(ingress)!: make tls section optional, no default cluster issuer
The template for Ingress always added the section for tls.
This is a problem if the target cluster does not need a tls configuration in the Ingress resource.

Now the template only adds the tls section if the value `ingress.tls` is set to true.

The template no longer adds the annotation `cert-manager.io/cluster-issuer`.