feat: support multiple user groups per user
Two new environment variables control if and how user groups in the user
auth token will be read and imported to the user in c4.

AUTH_USER_GROUPS_PROPERTY_NAME: Name of the property in the user info
object that contains the list of user group ids. Default: groups

AUTH_USE_USER_GROUPS_FROM_AUTH_PROVIDER: Controls if the groups from the
user info are imported into the user's group list. Default: false
The user's groups are replaced, not merged.

The import of groups is done in the auth service.

Refs: 316