tests: remove unnecessary cleanup and test steps (#644)
* tests: remove unnecessary cleanup step

since all assistants have unique names the cleanup should not be necessary

existing buckets will not be created a second time

this is a step on the way to parallel test execution

* tests: duplicate active conversation instead of latest to avoid races

* tests: do not wait unneccessarily for model tests

We will test the "test"button once and assume that it works for all other usages of the helper function.
This should shave seconds off of each test and reduce token consumption.

* docs: add comment

* tests: handle the unique names centrally

* tests: deterministic conversation titles