fix(REIS): do not process all file types with a detour over pdf
This comes with a trade off for the pdf viewer in the frontend. Until now, every file type could jump to the correct page in the preview pdf. But the price to pay was that every type was converted to pdf and then parsed and chunked, which leads to worse results and in most cases longer processing times.

Now we generate a pdf for each format, but parse the original -- with the exception of the office file formats. This way we can show the preview pdf, but can not jump to the page.

Also this PR:

* simplifies the code somewhat
* clarifies the names of some propoerties
* does microoptimization, which avouids conversion of the same file multiple times
* fixes a bug that chunk?siye was ignored for office formats