fix: avoid error when user groups from oidc are empty/undefined
If the switch AUTH_USE_USER_GROUPS_FROM_AUTH_PROVIDER is set, only the value 'true' activates the feature.

If the feature is activated but the token from the OIDC provider does not contain the group property or the groups are empty, the SQL query to fetch the groups failed.
Now the query is only executed if the groups are not empty.

If the groups from the OIDC token are empty, the user will not have any groups.

Refs: 692